+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/attestation/AggregatingAttestationPool.java
@@ -68,23 +68,13 @@ public class AggregatingAttestationPool implements SlotEventsChannel {
   /**
    * Default maximum number of attestations to store in the pool.
    *
-   * <p>With 1.2 million active validators, we'd expect around 37_500 attestations per slot; so 3
-   * slots worth of attestations is almost 120_000.
-   *
-   * <p>128 attestations perfectly packed at a 1.2 million validator set would be 1_200_000 / 32 /
-   * 64 bits, about 584 bits per aggregate. 128 of those is 74752 attestations if perfectly packed.
-   * Technically if we did have to cache 2 full slots of information, that would be roughly 150k
-   * cache size.
-   *
-   * <p>Because the real world exists, it's fair to expect that it's not all perfect, and 120k
-   * should be an adequately large cache to store current attestations plus some old ones that may
-   * not have been included so that we have plenty to choose if block building based on an expected
-   * 1.2 million validators.
+   * <p>With 2 million active validators, we'd expect around 62_500 attestations per slot; so 3
+   * slots worth of attestations is almost 187_500.
    *
    * <p>Strictly to cache all attestations for a full 2 epochs is significantly larger than this
    * cache.
    */
-  public static final int DEFAULT_MAXIMUM_ATTESTATION_COUNT = 120_000;
+  public static final int DEFAULT_MAXIMUM_ATTESTATION_COUNT = 187_500;
 
   private final Map<Bytes, MatchingDataAttestationGroup> attestationGroupByDataHash =
       new HashMap<>();